Update migrations to work with empty db

+++ site/gatsby-site/migrations/2022.12.16T22.18.54.remove-resources.js
@@ -2,9 +2,20 @@ const config = require('../config');
 
 /** @type {import('umzug').MigrationFn<any>} */
 exports.up = async ({ context: { client } }) => {
-  const resourcesCollection = client.db(config.realm.production_db.db_name).collection('resources');
+  const collections = await client
+    .db(config.realm.production_db.db_name)
+    .listCollections({ name: 'resources' })
+    .toArray();
 
-  await resourcesCollection.drop();
+  if (collections.length === 0) {
+    console.log('Collection resources does not exist.');
+  } else {
+    const resourcesCollection = client
+      .db(config.realm.production_db.db_name)
+      .collection('resources');
+
+    await resourcesCollection.drop();
+  }
 
   const taxaCollection = client.db(config.realm.production_db.db_name).collection('taxa');
